Swapped main/master, added 4/5 error code handling
Swapped main.py and master.py since repl.it always runs main.py first, so now repl.it can call fileConversion.py as well.
Added error handling for if the page is not found (ex. 404 error code), to output error to screen and .txt and terminate program.
Updated the output delimeter for Linux from a | to a ; to solve some issues.