Added description fetching
Now fetches the descriptions from the individual event pages.  Currently, there's an issue with some descriptions making two columns.
This also messes with Excel's manual import data from text file feature, as the headers are not recognized and must be manually moved.  Not sure how it works with the auto-.xlsx feature yet.