Big DTBville Merge update
The DTBville ICS scraper is now integrated with fayFlyScraper.
The dates and times for DTBville are fetched from each event's individual event page on the website because they are either wrong in the .ics file or I just can't figure out how to decode/parse them correctly.
Since DTBville scrapes each individual event's page, and currently there are over 400 events, the process takes a little over two minutes... A LONG TIME. Because of this, there is now a control file that has methods that print when each website's section begins and ends, as well as the duration of the process.  
The .ics function fetches the Location without the address after the @ symbol, the contact phone number and email (if it exists), as well as the cost.  If there is no cost, "na" is put in its place, but will probably be changed to put nothing.  

Still trying to figure out the description printing with its lovely encoding errors.

Conversion still isn't working on repl.it, but I haven't worked on it in a while.